requirements.txt
apiclient==1.0.4
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.26.0
google-api-python-client==1.7.9
google-auth==1.26.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.13.0
idna==2.8
mysql-client==0.0.1
mysql-connector==2.2.9
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pdfkit==0.6.1
protobuf==3.14.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.7
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
uritemplate==3.0.0
urllib3==1.25.3
